Update Amelia character and enhance game mechanics
Amelia's character data and related attributes have been significantly updated to improve her gaming experience. The CharacterValueExtension has been renamed for a clearer understanding. Amelia's ability details including name, description and cooldown time and passive descriptions have been updated for enhanced gameplay and understanding. Her health stats, movement speed and special values have been adjusted for better game balancing. The permanent upgrades type has been enriched by adding the 'stat type' attribute to several types ensuring variety and better gaming strategy. Changes have been made to the font asset as well for a smoother visual experience.

The changes were necessary in order to balance the game flow, make character abilities more understandable and interactive, and give players a better in-game experience.

+++ Assets/Objects/Players/PlayerStats.cs
@@ -242,7 +242,7 @@ namespace Objects.Players
 					AttackCount += (int) value;
 					break;
 				case PermUpgradeType.HealthRegen:
-					HealthRegen += (int) value;
+					HealthRegen += value;
 					break;
 				case PermUpgradeType.DamageIncrease:
 					DamagePercentageIncrease += value;
@@ -254,7 +254,7 @@ namespace Objects.Players
 					CooldownReductionPercentage += value;
 					break;
 				case PermUpgradeType.Health:
-					HealthMax += (int) value;
+					HealthMax += value;
 					break;
 				case PermUpgradeType.Curse:
 					EnemySpeedIncreasePercentage += value;
